Fixed data models enams. Fixed Authorization
