+++ CarShop/CarShop/CarShopClient/src/app/components/nav-bar/nav-bar/nav-bar.component.html
@@ -12,8 +12,6 @@
         </li>
       </ul>
       <ul class="nav nav-pills">
-        
-        
         <li class="nav-item">
           <div *ngIf="!loggedIn; else loggetOut">
             <span class="title-right">
