+++ CarShop/CarShop/CarShopClient/src/app/models/enums/EngineTypes.ts
@@ -1,0 +1,5 @@
+export enum EngineTypes {
+    PetrolEngine,
+    DieselEngine,
+    ElectroEngine
+}
