+++ CarShop/CarShop/CarShopClient/src/app/models/enums/OrderStatuses.ts
@@ -1,0 +1,4 @@
+export enum OrderStatuses {
+    InProgress,
+    Paid
+}
