+++ CarShop/CarShop/CarShopClient/src/app/models/enums/UserRoles.ts
@@ -1,0 +1,6 @@
+export enum UserRoles {
+    Admin = 1,
+    AdminAssistant,
+    AuthorizedUser,
+    NotAuthorizedUser
+}
