+++ CarShop/CarShop/CarShopClient/src/app/models/enums/WheelDrives.ts
@@ -1,0 +1,6 @@
+export enum WheelDrives {
+    FrontWheelDrive,
+    RealWheelDrive,
+    AllWheelDrive,
+    PlugInAllWheelDrive
+}
